backend/projects/requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
sqlparse==0.5.1
python-dotenv==1.0.1
mysqlclient==2.2.4